+++ pyoxidizer/src/pyrepackager/repackage.rs
@@ -401,6 +401,31 @@ fn filter_btreemap<V>(logger: &slog::Logger, m: &mut BTreeMap<String, V>, f: &BT
     }
 }
 
+lazy_static! {
+    static ref RE_CODING: regex::bytes::Regex =
+        { regex::bytes::Regex::new(r"^[ \t\f]*#.*?coding[:=][ \t]*([-_.a-zA-Z0-9]+)").unwrap() };
+}
+
+/// Derive the source encoding from Python source code.
+pub fn python_source_encoding(source: &[u8]) -> Vec<u8> {
+    // Default source encoding is UTF-8. But per PEP 263, the first or second
+    // line of source can match a regular expression to define a custom
+    // encoding.
+    let lines = source.split(|v| v == &b'\n');
+
+    for (i, line) in lines.enumerate() {
+        if i > 1 {
+            break;
+        }
+
+        if let Some(m) = RE_CODING.find(line) {
+            return m.as_bytes().to_vec();
+        }
+    }
+
+    b"utf-8".to_vec()
+}
+
 struct BytecodeRequest {
     source: Vec<u8>,
     optimize_level: i32,
@@ -829,6 +854,32 @@ pub fn resolve_python_resources(
             "removing extension module due to incompatibility: {}", e
         );
         embedded_extension_modules.remove(&String::from(*e));
+    }
+
+    // Audit Python source for __file__, which could be problematic.
+    let mut file_seen = false;
+
+    for (name, request) in &embedded_bytecode_requests {
+        // We can't just look for b"__file__ because the source file may be in
+        // encodings like UTF-16. So we need to decode to Unicode first then look for
+        // the code points.
+        let encoding = python_source_encoding(&request.source);
+
+        let encoder = match encoding_rs::Encoding::for_label(&encoding) {
+            Some(encoder) => encoder,
+            None => encoding_rs::UTF_8,
+        };
+
+        let (source, ..) = encoder.decode(&request.source);
+
+        if source.contains("__file__") {
+            warn!(logger, "warning: {} contains __file__", name);
+            file_seen = true;
+        }
+    }
+
+    if file_seen {
+        warn!(logger, "__file__ was encountered in some modules; PyOxidizer does not set __file__ and this may create problems at run-time; see https://github.com/indygreg/PyOxidizer/issues/69 for more");
     }
 
     let mut embedded_bytecodes: BTreeMap<String, PackagedModuleBytecode> = BTreeMap::new();